refactored routes file structure
