+++ README.md
@@ -1,5 +1,7 @@
 The backend that powers the HACS website!
 
+Following structure set by https://mean.io by Linnovate.
+
 Our current API Routes:
 
 - /officers
